Move enum definitions "no duplicates" test to correct file
